pyproject.toml
[tool.poetry]
name = "how-to-test"
version = "0.1.0"
description = ""
authors = ["Jared Dantis <jareddantis@gmail.com>"]
readme = "README.md"
packages = [{include = "backend"}]

[tool.poetry.dependencies]
python = "^3.13"
fastapi = {extras = ["standard"], version = "^0.115.8"}
starlette-context = "^0.3.6"
requests = "^2.32.3"

[tool.poetry.group.dev.dependencies]
ruff = "^0.9.6"
pytest = "^8.3.4"
playwright = "^1.50.0"
pytest-playwright = "^0.7.0"
types-requests = "^2.32.0.20241016"

[tool.ruff]
indent-width = 2

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I", "PL"]
ignore = [
  "PLR0913", # Too many arguments
]

[tool.ruff.format]
quote-style = "single"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
